Nutribot/requirements.txt
aiohttp==3.12.13
bcrypt==4.0.1
charset-normalizer==3.4.2
clarifai==10.11.0
decorator==5.2.1
googleapis-common-protos==1.70.0
httpcore==1.0.9
httpx==0.27
openai==1.52.2
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
minio==7.2.15
nest-asyncio==1.6.0
numpy==1.26.4
pandas==2.2.3
Pillow==10.1.0
platformdirs==4.3.8
playwright==1.49.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-telegram-bot==22.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
soundfile==0.13.1
supabase==2.10.0
replicate==1.0.7
requests==2.32.4
torch==2.2.0
gunicorn==23.0.0
Flask==3.1.1
xgboost==3.0.3